pivotal-181943101 display global color tokens, rebuild tokens, move stuff around
